+++ src/xchain/modules/folks-loan.ts
@@ -62,7 +62,6 @@ import type {
   PrepareRepayCall,
   PrepareRepayWithCollateralCall,
   PrepareSwitchBorrowTypeCall,
-  PrepareUpdateUserLoanPoolPoints,
   PrepareWithdrawCall,
 } from "../../common/types/module.js";
 import type { FolksTokenId } from "../../common/types/token.js";
@@ -868,24 +867,6 @@ export const prepare = {
       hubChain,
     );
   },
-
-  async updateUserLoanPoolPoints(loanIds: Array<LoanId>): Promise<PrepareUpdateUserLoanPoolPoints> {
-    const folksChain = FolksCore.getSelectedFolksChain();
-    assertHubChainSelected(folksChain.folksChainId, folksChain.network);
-    const hubChain = getHubChain(folksChain.network);
-
-    const userAddress = getSignerGenericAddress({
-      signer: FolksCore.getFolksSigner().signer,
-      chainType: folksChain.chainType,
-    });
-
-    return await FolksHubLoan.prepare.updateUserLoanPoolPoints(
-      FolksCore.getProvider<ChainType.EVM>(folksChain.folksChainId),
-      convertFromGenericAddress(userAddress, folksChain.chainType),
-      loanIds,
-      hubChain,
-    );
-  },
 };
 
 export const write = {
@@ -1146,18 +1127,6 @@ export const write = {
       FolksCore.getProvider<ChainType.EVM>(folksChain.folksChainId),
       FolksCore.getSigner<ChainType.EVM>(),
       accountId,
-      prepareCall,
-    );
-  },
-
-  async updateUserLoanPoolPoints(loanIds: Array<LoanId>, prepareCall: PrepareUpdateUserLoanPoolPoints) {
-    const folksChain = FolksCore.getSelectedFolksChain();
-    assertHubChainSelected(folksChain.folksChainId, folksChain.network);
-
-    return await FolksHubLoan.write.updateUserLoanPoolPoints(
-      FolksCore.getProvider<ChainType.EVM>(folksChain.folksChainId),
-      FolksCore.getSigner<ChainType.EVM>(),
-      loanIds,
       prepareCall,
     );
   },
