+++ src/xchain/modules/folks-rewards.ts
@@ -5,8 +5,9 @@ import { convertFromGenericAddress } from "../../common/utils/address.js";
 import { assertHubChainSelected, getSignerGenericAddress } from "../../common/utils/chain.js";
 import { FolksCore } from "../core/folks-core.js";
 
+import type { PrepareUpdateUserPointsInLoans } from "../../chains/evm/common/types/index.js";
 import type { LoanTypeInfo } from "../../chains/evm/hub/types/loan.js";
-import type { ActiveEpochs, UserPoints } from "../../chains/evm/hub/types/rewards.js";
+import type { ActiveEpochs, UserPoolPoints } from "../../chains/evm/hub/types/rewards.js";
 import type { ChainType } from "../../common/types/chain.js";
 import type { AccountId, LoanId } from "../../common/types/lending.js";
 import type {
@@ -16,6 +17,24 @@ import type {
 } from "../../common/types/module.js";
 
 export const prepare = {
+  async updateUserPointsInLoans(loanIds: Array<LoanId>): Promise<PrepareUpdateUserPointsInLoans> {
+    const folksChain = FolksCore.getSelectedFolksChain();
+    assertHubChainSelected(folksChain.folksChainId, folksChain.network);
+    const hubChain = getHubChain(folksChain.network);
+
+    const userAddress = getSignerGenericAddress({
+      signer: FolksCore.getFolksSigner().signer,
+      chainType: folksChain.chainType,
+    });
+
+    return await FolksHubRewards.prepare.updateUserPointsInLoans(
+      FolksCore.getProvider<ChainType.EVM>(folksChain.folksChainId),
+      convertFromGenericAddress(userAddress, folksChain.chainType),
+      loanIds,
+      hubChain,
+    );
+  },
+
   async updateAccountsPointsForRewards(
     accountIds: Array<AccountId>,
     activeEpochs: ActiveEpochs,
@@ -59,6 +78,18 @@ export const prepare = {
 };
 
 export const write = {
+  async updateUserPointsInLoans(loanIds: Array<LoanId>, prepareCall: PrepareUpdateUserPointsInLoans) {
+    const folksChain = FolksCore.getSelectedFolksChain();
+    assertHubChainSelected(folksChain.folksChainId, folksChain.network);
+
+    return await FolksHubRewards.write.updateUserPointsInLoans(
+      FolksCore.getProvider<ChainType.EVM>(folksChain.folksChainId),
+      FolksCore.getSigner<ChainType.EVM>(),
+      loanIds,
+      prepareCall,
+    );
+  },
+
   async updateAccountsPointsForRewards(
     accountIds: Array<AccountId>,
     prepareCall: PrepareUpdateAccountsPointsForRewardsCall,
@@ -98,12 +129,12 @@ export const read = {
     return getUnclaimedRewards(FolksCore.getHubProvider(), FolksCore.getSelectedNetwork(), accountId, activeEpochs);
   },
 
-  async userPoints(
+  async userPoolPoints(
     accountId: AccountId,
     loanIds: Array<LoanId>,
     loanTypesInfo: Partial<Record<LoanTypeId, LoanTypeInfo>>,
-  ): Promise<UserPoints> {
-    return FolksHubRewards.getUserPoints(
+  ): Promise<UserPoolPoints> {
+    return FolksHubRewards.getUserPoolPoints(
       FolksCore.getHubProvider(),
       FolksCore.getSelectedNetwork(),
       accountId,
